Move column encoder 2 into separate file
